rename tableheader dir to camelcase
