+++ src/frontend/src/views/Data/tableHeaders/commonSampleHeaders.tsx
