+++ src/frontend/src/views/Data/tableHeaders/sampleHeadersConfig.tsx
